queue/requirements.txt
aiohttp==3.8.3 ; python_version >= "3.9" and python_version < "4.0"
aiosignal==1.2.0 ; python_version >= "3.9" and python_version < "4.0"
aocref==2.0.6 ; python_version >= "3.9" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.9" and python_version < "4.0"
attrs==22.1.0 ; python_version >= "3.9" and python_version < "4.0"
cachetools==5.2.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2022.9.14 ; python_version >= "3.9" and python_version < "4"
charset-normalizer==2.1.1 ; python_version >= "3.9" and python_version < "4"
colorama==0.4.5 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
construct==2.8.16 ; python_version >= "3.9" and python_version < "4.0"
flatbuffers==2.0.7 ; python_version >= "3.9" and python_version < "4.0"
frozenlist==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
google-api-core==2.10.1 ; python_version >= "3.9" and python_version < "4.0"
google-api-core[grpc]==2.10.1 ; python_version >= "3.9" and python_version < "4.0"
google-auth==2.11.1 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-core==2.3.2 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-pubsub==2.13.6 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-storage==2.5.0 ; python_version >= "3.9" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
google-resumable-media==2.3.3 ; python_version >= "3.9" and python_version < "4.0"
googleapis-common-protos==1.56.4 ; python_version >= "3.9" and python_version < "4.0"
googleapis-common-protos[grpc]==1.56.4 ; python_version >= "3.9" and python_version < "4.0"
grpc-google-iam-v1==0.12.4 ; python_version >= "3.9" and python_version < "4.0"
grpcio-status==1.49.1 ; python_version >= "3.9" and python_version < "4.0"
grpcio==1.49.1 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4"
mgz @ git+https://github.com/happyleavesaoc/aoc-mgz@master ; python_version >= "3.9" and python_version < "4.0"
multidict==6.0.2 ; python_version >= "3.9" and python_version < "4.0"
proto-plus==1.22.1 ; python_version >= "3.9" and python_version < "4.0"
protobuf==4.21.6 ; python_version >= "3.9" and python_version < "4.0"
pyasn1-modules==0.2.8 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.9" and python_version < "4.0"
pycountry==22.3.5 ; python_version >= "3.9" and python_version < "4"
requests==2.28.1 ; python_version >= "3.9" and python_version < "4"
rsa==4.9 ; python_version >= "3.9" and python_version < "4"
ruamel-yaml-clib==0.2.6 ; platform_python_implementation == "CPython" and python_version < "3.11" and python_version >= "3.9"
ruamel-yaml==0.17.21 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
tabulate==0.8.10 ; python_version >= "3.9" and python_version < "4.0"
tqdm==4.64.1 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.12 ; python_version >= "3.9" and python_version < "4"
yarl==1.8.1 ; python_version >= "3.9" and python_version < "4.0"
